+++ src/components/organisms/StickyNote.tsx
@@ -1,12 +1,30 @@
+// @ts-nocheck
 import React, { useState, useContext, useRef, useEffect, useMemo } from "react";
 import { withStyles } from "@material-ui/core/styles";
 import { yellow, green, red, blue } from "@material-ui/core/colors";
-import { CanvasContext } from "../../contexts/CanvasContext";
+import { CanvasMaterialsContext } from "../../contexts/CanvasMaterialsContext";
 import { TextField, Menu, MenuItem, Radio, Divider } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 import classNames from "classnames";
 import { Rnd } from "react-rnd";
 import LockButton from "../atoms/LockButton";
+import { atom } from "recoil";
+
+export const stickyNoteState = atom({
+  key: "stickyNoteState",
+  default: {
+    color: "",
+    createdBy: "",
+    height: 0,
+    width: 0,
+    id: "",
+    isEdit: false,
+    word: "",
+    zIndex: 0,
+    positionX: 0,
+    positionY: 0,
+  },
+});
 type Props = {
   canvasId: string;
   setIsAreaClicked: (isAreaClicked: boolean) => void;
@@ -100,7 +118,7 @@ const StickyNote: React.FC<any> = (props) => {
     sendBackward,
     bringToFront,
     sendToBack,
-  } = useContext(CanvasContext);
+  } = useContext(CanvasMaterialsContext);
   const [mouseState, setMouseState] = useState<{
     mouseX: number;
     mouseY: number;
